requirements.txt
pandas
matplotlib
seaborn
numpy
scipy
jupyter
umap
sklearn.impute
sklearn.manifold
